requirements.txt
fastapi[standard]
python-jose
sqlalchemy
pyodbc
requests